requirements.txt
asgiref==3.3.1
beautifulsoup4==4.9.3
blinker==1.4
boto3==1.16.51
botocore==1.19.51
brotlipy==0.7.0
cssselect==1.1.0
defusedxml==0.6.0
dj-database-url==0.5.0
django-allauth==0.41.0
django-bootstrap4==2.3.1
django-braces==1.14.0
django-el-pagination==3.3.0
django-environ==0.4.5
django-heroku==0.3.1
django-model-utils==4.1.1
django-notifications-hq==1.6.0
django-recaptcha==2.0.6
django-social-share==2.1.1
django-storages==1.11.1
django-taggit==1.3.0
django-tinymce==3.2.0
django-widget-tweaks==1.4.8
Faker==4.1.0
idna==2.9
importlib-metadata==2.1.1
jmespath==0.10.0
jsonfield==3.1.0
lxml==4.6.2
markdown2==2.3.10
oauthlib==3.0.1
olefile==0.46
Pillow==7.1.2
psycopg2==2.8.6
PyJWT==1.7.1
pyOpenSSL==19.1.0
pyquery==1.4.3
python-dateutil==2.8.1
python3-openid==3.1.0
pytz==2020.1
readtime==1.1.1
requests-oauthlib==1.2.0
s3transfer==0.3.3
six==1.15.0
soupsieve==2.1
sqlparse==0.3.1
swapper==1.1.2.post1
text-unidecode==1.3
urllib3==1.25.9
whitenoise==5.2.0
zipp==3.4.0
